+++ Pupa/Sources/PupaApp/MyApps/FashionCompanionExample.swift
@@ -18,7 +18,7 @@ enum FashionCompanionExample: ExampleMyApp {
     }
 
     @MainActor
-    static func seedAgentsMd(globalMemory: MemoryStore, appRootOverride: URL? = nil) {
+    static func seedAgentsMd(globalMemory: MemoryStore?, appRootOverride: URL? = nil) {
         let appRoot = appRootOverride ?? MemoryStore.appRoot(myAppName: name)
         let appMemory = MemoryStore(rootOverride: appRoot)
         var wroteAny = false
@@ -33,7 +33,7 @@ enum FashionCompanionExample: ExampleMyApp {
                 wroteAny = true
             }
         }
-        if wroteAny { globalMemory.rescan() }
+        if wroteAny { globalMemory?.rescan() }
     }
 
     // MARK: - Builder