+++ Pupa/Sources/PupaApp/MyApps/ResearchTrackerExample.swift
@@ -20,7 +20,7 @@ enum ResearchTrackerExample: ExampleMyApp {
     }
 
     @MainActor
-    static func seedAgentsMd(globalMemory: MemoryStore, appRootOverride: URL? = nil) {
+    static func seedAgentsMd(globalMemory: MemoryStore?, appRootOverride: URL? = nil) {
         let appRoot = appRootOverride ?? MemoryStore.appRoot(myAppName: name)
         let appMemory = MemoryStore(rootOverride: appRoot)
         var wroteAny = false
@@ -35,7 +35,7 @@ enum ResearchTrackerExample: ExampleMyApp {
                 wroteAny = true
             }
         }
-        if wroteAny { globalMemory.rescan() }
+        if wroteAny { globalMemory?.rescan() }
     }
 
     // MARK: - Builder